common code: Sample rate conversion now supports any rate above 8000 Hz

+++ src/common/samplerate.h
@@ -1,10 +1,10 @@
 #include "filter.h"
 
 typedef struct samplerate {
-	int factor;
+	double factor;
 	struct {
 		double sum;
-		int sum_count;
+		double sum_count;
 		biquad_low_pass_t bq;
 	} down;
 	struct {
@@ -13,6 +13,6 @@ typedef struct samplerate {
 	} up;
 } samplerate_t;
 
-int init_samplerate(samplerate_t *state, int samplerate);
+int init_samplerate(samplerate_t *state, double samplerate);
 int samplerate_downsample(samplerate_t *state, int16_t *input, int input_num, int16_t *output);
 int samplerate_upsample(samplerate_t *state, int16_t *input, int input_num, int16_t *output);
